pyproject.toml
[project]
name = "ai-coding-pipeline"
version = "0.1.0"
description = "AI Coding 开发流水线 - 基于飞书"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "lark-oapi>=1.5.0",
    "litellm>=1.40.0",
    "PyGithub>=2.3.0",
    "celery[redis]>=5.4.0",
    "pydantic-settings>=2.5.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100
